Speed up spacemacs load time
Spacemacs layers are starting a bash shell to retrieve path information
during startup. The normal load time for the shell is very slow due to
things not relevent to path information. Force the library starting the
shell to load the bashrc file first. This will set a variable that will
prevent expensive operations from occuring.

With the reduction in things loading startup time is decreased from ~30
seconds to ~5 seconds.